Started a sketch for add new books in the array
Wrote some inputs form to obtain some values (meant as "title", "author" etc..)
for the books and a submit button.
Worked on the script side with a constructor for the books
Some functions to get new books
Created an array "myLibrary" to store all the books objects
Added a getRandomBookID fun to put an unique ID to every book
Added a "event.preventDefault()" for avoid the default work of the submit button.